refactor: rename OzmaModePlugin->DefaultModePlugin, ozma.rs->app_mode.rs
