+++ src/tmux.rs
@@ -14,7 +14,7 @@ mod webview_tokens;
 mod window_bar;
 mod window_bar_input;
 
-use crate::ozma::AppMode;
+use crate::app_mode::AppMode;
 use bevy::prelude::*;
 use copy_mode::CopyModePlugin;
 use dialog::DialogPlugin;
